Handling missing observations - resolved
Finished adding capability to handle missing observations (in the state variables) appropriately. Followed Eq. 3.4.75 f.f. of Harvey text. Key changes since last commit were to make sure that the dimensions work correctly in the updating equations (i.e. indexing things like the F and P matrices so that dimensions work correctly). Checked that this works if either one or both of the observations at a given time point are missing.

I have worked through the logic of the indexed math for the updating equations and I think that I did everything correctly as suggested by Harvey, but it probably wouldn't hurt to double-check this.